Support Android (ARMv7a/8a) via NDK
Adds new architectures armv7-a and armv8-a to the Makefile. Changes in the Makefile should not interfere with the remaining code and a single line in the source code is adapted for compatibility.

Compiling binaries for the Android platform is now very straight forward. Here is an example for the armv8-a architecture:
1.) Download NDK from here: https://developer.android.com/ndk/downloads/index.html
2.) Unzip NDK, for example to the directory ~/android-ndk-r21
3.) Enter the Honey /src directory and compile ARMv8-a binary:
     $ make build ARCH=armv8-a COMP=ndk NDK=~/android-ndk-r21